Refactor project structure. Set default peak and valley min distance to 1.
