+++ config/config.toml
@@ -1,12 +1,12 @@
 [first]
 # Min horizontal distance between peaks.
-thr_min_peak_horizontal_distance = 10_000
+thr_min_peak_horizontal_distance = 1
 
 # Min width of peak to consider.
 thr_min_peak_width = 20
 
 # Min horizontal distance between valleys.
-thr_min_valley_horizontal_distance = 10_000
+thr_min_valley_horizontal_distance = 1
 
 # Min width of valley to consider.
 thr_min_valley_width = 10
