+++ nucfreq/constants.py
@@ -1,0 +1,9 @@
+import re
+
+
+PLOT_FONT_SIZE = 16
+PLOT_HEIGHT = 6
+PLOT_WIDTH = 16
+PLOT_DPI = 600
+PLOT_YLIM = 100
+RGX_REGION = re.compile(r"(.+):(\d+)-(\d+)")
